[backend/frontend] refactor(HomePage): Simplify HomepageList (#2157)
