+++ apps/portal-api/src/modules/user-service/user-service.domain.test.ts
@@ -40,8 +40,6 @@ const SIMPLE = TEST_ORGANIZATIONS.SECOND_ORGANIZATION.USERS.SIMPLE;
 const ADMIN = TEST_ORGANIZATIONS.SECOND_ORGANIZATION.USERS.ADMIN_ORGA;
 const SECOND_ORG_ID = TEST_ORGANIZATIONS.SECOND_ORGANIZATION.ID;
 
-const DAY = 86_400_000;
-
 // ---------------------------------------------------------------------------
 // Shared DB helpers
 // ---------------------------------------------------------------------------
@@ -427,308 +425,6 @@ describe('userServiceDomain', () => {
         subscription_id: sub.id,
       });
       expect(rows).toHaveLength(1);
-    });
-  });
-
-  describe('loadUserServiceByUser', () => {
-    const createdSubscriptionIds: SubscriptionId[] = [];
-
-    const simpleUser = {
-      id: SIMPLE.ID,
-      selected_organization_id: SECOND_ORG_ID,
-    } as Parameters<typeof UserServiceDomain.loadUserServiceByUser>[0];
-
-    const defaultOpts = {
-      first: 50,
-      orderBy: 'User_Service.id',
-      orderMode: 'asc',
-    };
-
-    beforeEach(() => {
-      vi.spyOn(mailService, 'sendMail').mockResolvedValue(undefined);
-    });
-
-    afterEach(async () => {
-      for (const subId of createdSubscriptionIds) {
-        await cleanupUserServices(subId);
-        await TestHelper.subscription.delete({ id: subId });
-      }
-      createdSubscriptionIds.length = 0;
-    });
-
-    /** Creates an ACCEPTED subscription and enrols simpleUser into it. */
-    const createSubscriptionWithUser = async (overrides?: {
-      start_date?: Date;
-      end_date?: Date | null;
-      status?: string;
-      organization_id?: OrganizationId;
-      service_instance_id?: ServiceInstanceId;
-    }) => {
-      const subId = uuidv4() as SubscriptionId;
-      createdSubscriptionIds.push(subId);
-
-      await createSubscription({
-        id: subId,
-        service_instance_id:
-          overrides?.service_instance_id ?? SERVICES.INSTANCES.VAULT.ID,
-        organization_id: overrides?.organization_id ?? SECOND_ORG_ID,
-        start_date: overrides?.start_date ?? new Date(Date.now() - DAY),
-        end_date:
-          overrides?.end_date !== undefined
-            ? overrides.end_date
-            : new Date(Date.now() + DAY * 365),
-        billing: 0,
-        status: overrides?.status ?? SubscriptionStatus.ACCEPTED,
-      });
-
-      const subscription = await TestHelper.subscription.loadAll({ id: subId });
-
-      if (
-        (overrides?.status ?? SubscriptionStatus.ACCEPTED) ===
-        SubscriptionStatus.ACCEPTED
-      ) {
-        await insertUserService(simpleUser.id, subId);
-      }
-
-      return { subId, subscription };
-    };
-
-    it('should return one edge for a user enrolled in a single active subscription', async () => {
-      await createSubscriptionWithUser();
-
-      const result = await UserServiceDomain.loadUserServiceByUser(
-        simpleUser,
-        defaultOpts
-      );
-
-      expect(result.totalCount).toBe('1');
-      expect(result.edges).toHaveLength(1);
-      expect(result.edges[0]!.node!.user_id).toBe(simpleUser.id);
-    });
-
-    it('should return one edge per active subscription when enrolled in multiple services', async () => {
-      await createSubscriptionWithUser();
-
-      const subId2 = uuidv4() as SubscriptionId;
-      createdSubscriptionIds.push(subId2);
-      await createSubscription({
-        id: subId2,
-        service_instance_id: SERVICES.INSTANCES.OPENAEV_SCENARIOS.ID,
-        organization_id: SECOND_ORG_ID,
-        start_date: new Date(Date.now() - DAY),
-        end_date: new Date(Date.now() + DAY * 365),
-        billing: 0,
-        status: SubscriptionStatus.ACCEPTED,
-      });
-      await insertUserService(simpleUser.id, subId2);
-
-      const result = await UserServiceDomain.loadUserServiceByUser(
-        simpleUser,
-        defaultOpts
-      );
-
-      expect(result.totalCount).toBe('2');
-      expect(result.edges).toHaveLength(2);
-      expect(result.edges.every((e) => e.node!.user_id === simpleUser.id)).toBe(
-        true
-      );
-    });
-
-    it('should return an empty connection when the user has no UserService rows', async () => {
-      const result = await UserServiceDomain.loadUserServiceByUser(
-        simpleUser,
-        defaultOpts
-      );
-
-      expect(result.totalCount).toBe('0');
-      expect(result.edges).toHaveLength(0);
-    });
-
-    it('should return an empty connection for a non-existent user id', async () => {
-      const ghostUser = {
-        id: uuidv4(),
-        selected_organization_id: SECOND_ORG_ID,
-      } as Parameters<typeof UserServiceDomain.loadUserServiceByUser>[0];
-
-      const result = await UserServiceDomain.loadUserServiceByUser(
-        ghostUser,
-        defaultOpts
-      );
-
-      expect(result.totalCount).toBe('0');
-      expect(result.edges).toHaveLength(0);
-    });
-
-    it('should exclude subscriptions whose status is not ACCEPTED', async () => {
-      const subId = uuidv4() as SubscriptionId;
-      createdSubscriptionIds.push(subId);
-      await createSubscription({
-        id: subId,
-        service_instance_id: SERVICES.INSTANCES.VAULT.ID,
-        organization_id: SECOND_ORG_ID,
-        start_date: new Date(Date.now() - DAY),
-        end_date: new Date(Date.now() + DAY * 365),
-        billing: 0,
-        status: SubscriptionStatus.REQUESTED,
-      });
-      await TestHelper.user_Service.create({
-        user_id: simpleUser.id,
-        subscription_id: subId,
-      });
-
-      const result = await UserServiceDomain.loadUserServiceByUser(
-        simpleUser,
-        defaultOpts
-      );
-
-      expect(result.totalCount).toBe('0');
-      expect(result.edges).toHaveLength(0);
-    });
-
-    it('should exclude subscriptions whose end_date is in the past', async () => {
-      const subId = uuidv4() as SubscriptionId;
-      createdSubscriptionIds.push(subId);
-      await createSubscription({
-        id: subId,
-        service_instance_id: SERVICES.INSTANCES.VAULT.ID,
-        organization_id: SECOND_ORG_ID,
-        start_date: new Date(Date.now() - DAY * 30),
-        end_date: new Date(Date.now() - DAY),
-        billing: 0,
-        status: SubscriptionStatus.ACCEPTED,
-      });
-      await insertUserService(simpleUser.id, subId);
-
-      const result = await UserServiceDomain.loadUserServiceByUser(
-        simpleUser,
-        defaultOpts
-      );
-
-      expect(result.totalCount).toBe('0');
-      expect(result.edges).toHaveLength(0);
-    });
-
-    it('should include active subscriptions with no end_date (ongoing)', async () => {
-      await createSubscriptionWithUser({ end_date: null });
-
-      const result = await UserServiceDomain.loadUserServiceByUser(
-        simpleUser,
-        defaultOpts
-      );
-
-      expect(result.totalCount).toBe('1');
-      expect(result.edges).toHaveLength(1);
-    });
-
-    it('should exclude subscriptions whose start_date is in the future', async () => {
-      const subId = uuidv4() as SubscriptionId;
-      createdSubscriptionIds.push(subId);
-      await createSubscription({
-        id: subId,
-        service_instance_id: SERVICES.INSTANCES.VAULT.ID,
-        organization_id: SECOND_ORG_ID,
-        start_date: new Date(Date.now() + DAY),
-        end_date: new Date(Date.now() + DAY * 365),
-        billing: 0,
-        status: SubscriptionStatus.ACCEPTED,
-      });
-      await insertUserService(simpleUser.id, subId);
-
-      const result = await UserServiceDomain.loadUserServiceByUser(
-        simpleUser,
-        defaultOpts
-      );
-
-      expect(result.totalCount).toBe('0');
-      expect(result.edges).toHaveLength(0);
-    });
-
-    it('should exclude subscriptions belonging to a different organization', async () => {
-      const subId = uuidv4() as SubscriptionId;
-      createdSubscriptionIds.push(subId);
-      await createSubscription({
-        id: subId,
-        service_instance_id: SERVICES.INSTANCES.VAULT.ID,
-        organization_id: TEST_ORGANIZATIONS.FILIGRAN.ID,
-        start_date: new Date(Date.now() - DAY),
-        end_date: new Date(Date.now() + DAY * 365),
-        billing: 0,
-        status: SubscriptionStatus.ACCEPTED,
-      });
-      await insertUserService(simpleUser.id, subId);
-
-      const result = await UserServiceDomain.loadUserServiceByUser(
-        simpleUser,
-        defaultOpts
-      );
-
-      expect(result.totalCount).toBe('0');
-      expect(result.edges).toHaveLength(0);
-    });
-
-    it('should return correct pagination shape (edges, pageInfo, totalCount)', async () => {
-      await createSubscriptionWithUser();
-
-      const result = await UserServiceDomain.loadUserServiceByUser(
-        simpleUser,
-        defaultOpts
-      );
-
-      expect(typeof result.totalCount).toBe('string');
-      expect(Array.isArray(result.edges)).toBe(true);
-      expect(result.pageInfo).toBeDefined();
-      expect(typeof result.pageInfo.hasNextPage).toBe('boolean');
-      expect(typeof result.pageInfo.hasPreviousPage).toBe('boolean');
-
-      const edge = result.edges[0]!;
-      expect(typeof edge.cursor).toBe('string');
-      expect(edge.cursor.length).toBeGreaterThan(0);
-      expect(edge.node).toBeDefined();
-    });
-
-    it('should include service_name and ordering columns on each edge node', async () => {
-      await createSubscriptionWithUser();
-
-      const result = await UserServiceDomain.loadUserServiceByUser(
-        simpleUser,
-        defaultOpts
-      );
-
-      const node = result.edges[0]!.node as UserService & {
-        service_name: string | null;
-        ordering: number | null;
-      };
-
-      expect(Object.prototype.hasOwnProperty.call(node, 'service_name')).toBe(
-        true
-      );
-      expect(Object.prototype.hasOwnProperty.call(node, 'ordering')).toBe(true);
-    });
-
-    it('should only return services belonging to the queried user', async () => {
-      await createSubscriptionWithUser();
-
-      const adminSubId = uuidv4() as SubscriptionId;
-      createdSubscriptionIds.push(adminSubId);
-      await createSubscription({
-        id: adminSubId,
-        service_instance_id: SERVICES.INSTANCES.OPENAEV_SCENARIOS.ID,
-        organization_id: SECOND_ORG_ID,
-        start_date: new Date(Date.now() - DAY),
-        end_date: new Date(Date.now() + DAY * 365),
-        billing: 0,
-        status: SubscriptionStatus.ACCEPTED,
-      });
-      await insertUserService(ADMIN.ID, adminSubId);
-
-      const result = await UserServiceDomain.loadUserServiceByUser(
-        simpleUser,
-        defaultOpts
-      );
-
-      expect(result.totalCount).toBe('1');
-      expect(result.edges).toHaveLength(1);
-      expect(result.edges[0]!.node!.user_id).toBe(simpleUser.id);
     });
   });
