+++ apps/portal-front/src/components/service/service-list.tsx
@@ -1,150 +1,88 @@
 'use client';
 
-import useDecodedQuery from '@/hooks/useDecodedQuery';
+import { RegisterRegisteredPlatformsQuery } from '@/components/registration/register/register.graphql';
+import ServiceInstanceCard from '@/components/service/service-instance-card';
+import { useOrgaFreeTrial } from '@/components/service/trial-instances/useOrgaFreeTrials';
 import {
-  isAutomaticSubscriptionService,
-  isExternalService,
+  freeTrialSkeletonToServiceInstanceCardData,
+  publicServiceInstanceToInstanceCardData,
+  registeredPlatformToServiceInstanceCardData,
 } from '@/utils/services';
-import { toast } from '@filigran/ui';
-import { ServiceDefinitionIdentifierEnum } from '@generated/models/ServiceDefinitionIdentifier.enum';
-import { publicServiceQuery } from '@generated/publicServiceQuery.graphql';
+import { PlatformIdentifierEnum } from '@generated/models/PlatformIdentifier.enum';
+import registerRegisteredPlatformListFragmentGraphql, {
+  registerRegisteredPlatformListFragment$key,
+} from '@generated/registerRegisteredPlatformListFragment.graphql';
 import { registerRegisteredPlatformsQuery } from '@generated/registerRegisteredPlatformsQuery.graphql';
 import { serviceList_fragment$data } from '@generated/serviceList_fragment.graphql';
-import { userServiceOwnedQuery } from '@generated/userServiceOwnedQuery.graphql';
-import { userServicesOwned_fragment$data } from '@generated/userServicesOwned_fragment.graphql';
 import { useTranslations } from 'next-intl';
-import { useMemo } from 'react';
-import { PreloadedQuery } from 'react-relay';
-import AvailableServices from './home/available-services';
-import HighligthedServices from './home/highlighted-services';
-import { useServiceQueries } from './home/hooks/useServiceQueries';
-import OwnedServices from './home/owned-services';
+import {
+  PreloadedQuery,
+  usePreloadedQuery,
+  useRefetchableFragment,
+} from 'react-relay';
 
 interface ServiceProps {
-  queryRefUserServiceOwned: PreloadedQuery<userServiceOwnedQuery>;
-  queryRefServiceList: PreloadedQuery<publicServiceQuery>;
   queryRefRegisteredPlatforms: PreloadedQuery<registerRegisteredPlatformsQuery>;
-  onUpdate: () => void;
+  serviceData: serviceList_fragment$data[];
 }
 
 const ServiceList = ({
-  queryRefUserServiceOwned,
-  queryRefServiceList,
   queryRefRegisteredPlatforms,
-  onUpdate,
+  serviceData,
 }: ServiceProps) => {
   const t = useTranslations();
-  const { h } = useDecodedQuery();
+  const { availableTrials } = useOrgaFreeTrial();
 
-  const handleSuccess = (message: string) => {
-    toast({
-      title: t('Utils.Success'),
-      description: <>{message}</>,
-    });
+  const getRegisteredPlatforms = (
+    queryRef: PreloadedQuery<registerRegisteredPlatformsQuery>
+  ) => {
+    const queryData = usePreloadedQuery<registerRegisteredPlatformsQuery>(
+      RegisterRegisteredPlatformsQuery,
+      queryRef
+    );
+
+    const [data] = useRefetchableFragment<
+      registerRegisteredPlatformsQuery,
+      registerRegisteredPlatformListFragment$key
+    >(registerRegisteredPlatformListFragmentGraphql, queryData);
+    return data.registeredPlatforms ?? [];
   };
-  const handleError = (error: Error) => {
-    toast({
-      variant: 'destructive',
-      title: t('Utils.Error'),
-      description: <>{t(`Error.Server.${error.message}`)}</>,
-    });
-  };
 
-  if (h) {
-    toast({
-      description: <>{t(`Service.HighlightedServices`)}</>,
-    });
-  }
-
-  const {
-    ownedServices,
-    publicServices,
-    registeredPlatforms,
-    addSubscriptionInDb,
-  } = useServiceQueries(
-    queryRefUserServiceOwned,
-    queryRefServiceList,
-    queryRefRegisteredPlatforms,
-    onUpdate,
-    handleSuccess,
-    handleError
+  const registeredPlatforms = getRegisteredPlatforms(
+    queryRefRegisteredPlatforms
   );
 
-  const {
-    owned: publicOwnedServices,
-    notOwned: availableServices,
-    highlighted: publicHighlightedServices,
-  } = useMemo(() => {
-    return publicServices.reduce(
-      (acc, service) => {
-        if (service.service_definition?.identifier === h) {
-          acc.highlighted.push(service);
-        }
-        // A service is owned if it is subscribed or if it is an external service or an autoJoin service
-        else if (
-          service.organization_subscribed ||
-          isExternalService(
-            service.service_definition!
-              .identifier as ServiceDefinitionIdentifierEnum
-          ) ||
-          isAutomaticSubscriptionService(
-            service.service_definition!
-              .identifier as ServiceDefinitionIdentifierEnum
-          )
-        ) {
-          acc.owned.push(service);
-        } else {
-          acc.notOwned.push(service);
-        }
-        return acc;
-      },
-      {
-        owned: [] as serviceList_fragment$data[],
-        notOwned: [] as serviceList_fragment$data[],
-        highlighted: [] as serviceList_fragment$data[],
-      }
-    );
-  }, [publicServices, h]);
+  const freeTrialsSkeletonDataCards = availableTrials.map(
+    (platformIdentifier) =>
+      freeTrialSkeletonToServiceInstanceCardData(
+        platformIdentifier as PlatformIdentifierEnum,
+        t
+      )
+  );
 
-  const {
-    regular: regularOwnedServices,
-    highlighted: ownedHighlightedServices,
-  } = useMemo(() => {
-    return ownedServices.reduce(
-      (acc, service) => {
-        if (
-          service.subscription!.service_instance!.service_definition!
-            .identifier === h
-        ) {
-          acc.highlighted.push(service);
-        } else {
-          acc.regular.push(service);
-        }
-        return acc;
-      },
-      {
-        regular: [] as userServicesOwned_fragment$data[],
-        highlighted: [] as userServicesOwned_fragment$data[],
-      }
-    );
-  }, [ownedServices, h]);
+  const sortedServices = [
+    ...freeTrialsSkeletonDataCards,
+    ...registeredPlatforms.map((platform) =>
+      registeredPlatformToServiceInstanceCardData(platform, t)
+    ),
+    ...serviceData.map(publicServiceInstanceToInstanceCardData),
+  ].sort((a, b) => a!.ordering - b!.ordering);
 
   return (
     <>
-      <HighligthedServices
-        publicServices={publicHighlightedServices}
-        ownedServices={ownedHighlightedServices}
-        addSubscriptionInDb={addSubscriptionInDb}
-      />
-      <AvailableServices
-        services={availableServices}
-        addSubscriptionInDb={addSubscriptionInDb}
-      />
-      <OwnedServices
-        services={regularOwnedServices}
-        publicServices={publicOwnedServices}
-        registeredPlatforms={registeredPlatforms}
-      />
+      <ul
+        className={
+          'grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 2xl:grid-cols-4 gap-l'
+        }>
+        {sortedServices.map((service, index) => {
+          return (
+            <ServiceInstanceCard
+              key={`${service.id}-${index}`}
+              serviceInstance={service}
+            />
+          );
+        })}
+      </ul>
     </>
   );
 };
