--- apps/portal-front/src/components/service/service.types.ts
@@ -1,4 +1,0 @@
-import { publicServiceList_services$data } from '@generated/publicServiceList_services.graphql';
-
-export type PublicService =
-  publicServiceList_services$data['publicServiceInstances']['edges'][number]['node'];
