+++ apps/portal-front/src/components/service/user_service.graphql.ts
@@ -12,17 +12,6 @@ export const UserServiceFromSubscription = graphql`
   }
 `;
 
-export const UserServiceOwnedQuery = graphql`
-  query userServiceOwnedQuery(
-    $count: Int!
-    $cursor: ID
-    $orderBy: UserServiceOrdering!
-    $orderMode: OrderingMode!
-  ) {
-    ...userServiceOwnedUser
-  }
-`;
-
 export const userServiceFromSubscriptionFragment = graphql`
   fragment userServiceFromSubscription on Query
   @refetchable(queryName: "ServiceUserFromSubscriptionPaginationQuery") {
@@ -39,27 +28,6 @@ export const userServiceFromSubscriptionFragment = graphql`
         node {
           id
           ...userServices_fragment
-        }
-      }
-    }
-  }
-`;
-
-export const userServiceOwnedFragment = graphql`
-  fragment userServiceOwnedUser on Query
-  @refetchable(queryName: "ServiceUserOwnedPaginationQuery") {
-    userServiceOwned(
-      first: $count
-      after: $cursor
-      orderBy: $orderBy
-      orderMode: $orderMode
-    ) {
-      __id
-      totalCount
-      edges {
-        node {
-          id
-          ...userServicesOwned_fragment @relay(mask: false)
         }
       }
     }
@@ -87,31 +55,6 @@ export const userServiceFragment = graphql`
       first_name
       email
     }
-  }
-`;
-
-export const userServicesOwnedFragment = graphql`
-  fragment userServicesOwned_fragment on UserService {
-    id
-    user_service_capability {
-      generic_service_capability {
-        name
-      }
-      subscription_capability {
-        service_capability {
-          id
-          name
-        }
-      }
-    }
-    subscription {
-      id
-      status
-      service_instance {
-        ...serviceList_fragment @relay(mask: false)
-      }
-    }
-    ordering
   }
 `;
 
@@ -155,36 +98,6 @@ export const UserServiceCreateMutation = graphql`
   ) {
     addUserService(input: $input)
       @prependNode(connections: $connections, edgeTypeName: "UserServiceEdge") {
-      id
-      user {
-        id
-        first_name
-        last_name
-        email
-      }
-      user_service_capability {
-        id
-        generic_service_capability {
-          id
-          name
-        }
-        subscription_capability {
-          id
-          service_capability {
-            id
-            description
-            name
-          }
-        }
-      }
-    }
-  }
-`;
-export const UserServiceAddYourselfMutation = graphql`
-  mutation userServiceAddYourselfMutation(
-    $input: UserServiceAddYourselfInput!
-  ) {
-    addYourselfInUserService(input: $input) {
       id
       user {
         id
